+++ app/api/routes/models.py
@@ -8,10 +8,12 @@ from app.db.schemas.user import Permission
 logger = logging.getLogger(__name__)
 router = APIRouter(prefix="/models", tags=["Models"])
 
+CROP_MODEL_PATH = os.path.join(os.environ.get("MODELS_VOLUME_PATH"), "crop_model")
+
 
 @router.get("")
 async def list_models():
-    models = os.listdir(os.environ.get("MODELS_VOLUME_PATH"))
+    models = os.listdir(CROP_MODEL_PATH)
     # remove .pt extension
     models = [model[:-3] for model in models if model.endswith(".pt")]
     models_sorted = [models.pop(models.index("default"))] + sorted(models)
@@ -31,7 +33,7 @@ async def upload_model(file: UploadFile):
     # Save the uploaded model file
     if not file.filename.endswith(".pt"):
         raise HTTPException(400, "Invalid file")
-    model_path = os.path.join(os.environ.get("MODELS_VOLUME_PATH"), file.filename)
+    model_path = os.path.join(CROP_MODEL_PATH, file.filename)
     with open(model_path, "wb") as f:
         f.write(await file.read())
     logger.info(f"Uploaded model: {file.filename}")
@@ -47,7 +49,7 @@ async def upload_model(file: UploadFile):
     ],
 )
 async def delete_model(model_name: str):
-    model_path = os.path.join(os.environ.get("MODELS_VOLUME_PATH"), f"{model_name}.pt")
+    model_path = os.path.join(CROP_MODEL_PATH, f"{model_name}.pt")
     if not os.path.exists(model_path):
         raise HTTPException(404, "Model not found")
     os.remove(model_path)