+++ tests/integration/ndk.py
@@ -5,8 +5,7 @@ import requests
 API_URL = "http://localhost:8000"
 
 
-def main():
-    # Login as admin
+def login_for_token():
     response = requests.post(
         f"{API_URL}/users/login",
         data={
@@ -16,6 +15,12 @@ def main():
     )
     response.raise_for_status()
     token = response.json()["access_token"]
+    return token
+
+
+def main():
+    # Login as admin
+    token = login_for_token()
 
     # Use the token for subsequent requests
     bearer_headers = {"Authorization": f"Bearer {token}"}
@@ -124,7 +129,7 @@ def main():
         response.raise_for_status()
 
         assert response.json()["state"] == "retrain"
-        files = os.listdir(f"{os.environ['RETRAIN_VOLUME_PATH']}/{title_id}")
+        files = os.listdir(f"{os.environ['SCANS_VOLUME_PATH']}/{title_id}")
         assert set(files) == {
             "1-a0001.jpg",
             "1-a0002.jpg",
@@ -167,6 +172,11 @@ def main():
     response = requests.delete(f"{API_URL}/groups/{group_id}", headers=bearer_headers)
     response.raise_for_status()
     print("Cleaned up created group and title.")
+
+    scan_directory = f"{os.environ['SCANS_VOLUME_PATH']}/{title_id}"
+    assert not os.path.exists(scan_directory), (
+        f"Scan directory {scan_directory} still exists after cleanup."
+    )
 
     return True